add: getPostData
